BETA1.18 Fixed the credits and music manager.
The Credits scene is implemented, placeholder for image and mock-button in menu to get to the credits scene. Press A or start to get back from the credits scene.
The music in main001 now is randomized between the music that the designers supplied. If any more music tracks get added, you can change the array size on the music manager.