BETA1.19
Added:
Randomised level when the play button is pressed.

Modified:
Changed all of the LoadLevel functions to use the string name of the scenes instead of the number in the build settings.

Fixed:
The build settings scenes were all messed up so I have fixed them and added the missing ones.